+++ ydb/mvp/oidc_proxy/oidc_protected_page_handler.cpp
@@ -1,0 +1,23 @@
+#include "oidc_protected_page_handler.h"
+#include "oidc_protected_page_nebius.h"
+#include "oidc_protected_page_yandex.h"
+
+namespace NMVP {
+TProtectedPageHandler::TProtectedPageHandler(const NActors::TActorId& httpProxyId, const TOpenIdConnectSettings& settings)
+    : TBase(&TProtectedPageHandler::StateWork)
+    , HttpProxyId(httpProxyId)
+    , Settings(settings)
+{}
+
+void TProtectedPageHandler::Handle(NHttp::TEvHttpProxy::TEvHttpIncomingRequest::TPtr event, const NActors::TActorContext& ctx) {
+    switch (Settings.AuthProfile) {
+        case NMVP::EAuthProfile::Yandex:
+            ctx.Register(new THandlerSessionServiceCheckYandex(event->Sender, event->Get()->Request, HttpProxyId, Settings));
+            break;
+        case NMVP::EAuthProfile::Nebius:
+            ctx.Register(new THandlerSessionServiceCheckNebius(event->Sender, event->Get()->Request, HttpProxyId, Settings));
+            break;
+    }
+}
+
+}  // NMVP
