+++ ydb/mvp/oidc_proxy/oidc_session_create_nebius.h
@@ -10,9 +10,9 @@ private:
 
 public:
     THandlerSessionCreateNebius(const NActors::TActorId& sender,
-                          const NHttp::THttpIncomingRequestPtr& request,
-                          const NActors::TActorId& httpProxyId,
-                          const TOpenIdConnectSettings& settings)
+                                const NHttp::THttpIncomingRequestPtr& request,
+                                const NActors::TActorId& httpProxyId,
+                                const TOpenIdConnectSettings& settings)
         : THandlerSessionCreate(sender, request, httpProxyId, settings)
         {}
